page title component & repository,suggestion component add #13
