--- src/components/SearchRepository.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const SearchRepository: React.FC = () => {
-    return (
-        <div>SearchRepository</div>
-    )
-}
-
-export default SearchRepository
